Minor code optimization
